+++ nessi/modbuilder/interp2d/__init__.py
@@ -23,12 +23,7 @@ from __future__ import print_function
 # Import nessi.grd classes and functions
 from .grdwrap import voronoi
 from .grdwrap import idweight
-from .grdwrap import sibson1
-from .grdwrap import sibson2
-
-from .grdinterp import grdvoronoi 
-from .grdinterp import grdinvdist
-from .grdinterp import grdsibson
+from .grdwrap import sibson
 
 if __name__ == '__main__':
     import doctest